+++ cmd/server/gzip.go
@@ -1,74 +1,69 @@
 package main
 
-import (
-	"compress/gzip"
-	"fmt"
-	"io"
-	"net/http"
-)
+// import (
+// 	"compress/gzip"
+// 	"fmt"
+// 	"io"
+// 	"net/http"
+// )
 
-// compressWriter реализует интерфейс http.ResponseWriter и позволяет прозрачно для сервера
-// сжимать передаваемые данные и выставлять правильные HTTP-заголовки
-type compressWriter struct {
-	w  http.ResponseWriter
-	zw *gzip.Writer
-}
+// type compressWriter struct {
+// 	w  http.ResponseWriter
+// 	zw *gzip.Writer
+// }
 
-func newCompressWriter(w http.ResponseWriter) *compressWriter {
-	return &compressWriter{
-		w:  w,
-		zw: gzip.NewWriter(w),
-	}
-}
+// func newCompressWriter(w http.ResponseWriter) *compressWriter {
+// 	return &compressWriter{
+// 		w:  w,
+// 		zw: gzip.NewWriter(w),
+// 	}
+// }
 
-func (c *compressWriter) Header() http.Header {
-	return c.w.Header()
-}
+// func (c *compressWriter) Header() http.Header {
+// 	return c.w.Header()
+// }
 
-func (c *compressWriter) Write(p []byte) (int, error) {
-	return c.zw.Write(p)
-}
+// func (c *compressWriter) Write(p []byte) (int, error) {
+// 	return c.zw.Write(p)
+// }
 
-func (c *compressWriter) WriteHeader(statusCode int) {
-	if statusCode < 300 {
-		c.w.Header().Set("Content-Encoding", "gzip")
-	}
-	c.w.WriteHeader(statusCode)
-}
+// func (c *compressWriter) WriteHeader(statusCode int) {
+// 	if statusCode < 300 {
+// 		c.w.Header().Set("Content-Encoding", "gzip")
+// 	}
+// 	c.w.WriteHeader(statusCode)
+// }
 
-// Close закрывает gzip.Writer и досылает все данные из буфера.
-func (c *compressWriter) Close() error {
-	return c.zw.Close()
-}
+// func (c *compressWriter) Close() error {
+// 	return c.zw.Close()
+// }
 
-// compressReader реализует интерфейс io.ReadCloser и позволяет прозрачно для сервера
-// декомпрессировать получаемые от клиента данные
-type compressReader struct {
-	r  io.ReadCloser
-	zr *gzip.Reader
-}
+// type compressReader struct {
+// 	r  io.ReadCloser
+// 	zr *gzip.Reader
+// }
 
-func newCompressReader(r io.ReadCloser) (*compressReader, error) {
-	zr, err := gzip.NewReader(r)
-	if err != nil {
-		err = fmt.Errorf("ошибка 1: %s", err)
-		return nil, err
-	}
+// func newCompressReader(r io.ReadCloser) (*compressReader, error) {
+// 	zr, err := gzip.NewReader(r)
+// 	if err != nil {
+// 		err = fmt.Errorf("ошибка 1: %s", err)
+// 		return nil, err
+// 	}
 
-	return &compressReader{
-		r:  r,
-		zr: zr,
-	}, nil
-}
+// 	return &compressReader{
+// 		r:  r,
+// 		zr: zr,
+// 	}, nil
+// }
 
-func (c compressReader) Read(p []byte) (n int, err error) {
-	return c.zr.Read(p)
-}
+// func (c compressReader) Read(p []byte) (n int, err error) {
+// 	return c.zr.Read(p)
+// }
 
-func (c *compressReader) Close() error {
-	if err := c.r.Close(); err != nil {
-		err = fmt.Errorf("ошибка 2: %s", err)
-		return err
-	}
-	return c.zr.Close()
-}
+// func (c *compressReader) Close() error {
+// 	if err := c.r.Close(); err != nil {
+// 		err = fmt.Errorf("ошибка 2: %s", err)
+// 		return err
+// 	}
+// 	return c.zr.Close()
+// }
