+++ internal/constants/constants.go
@@ -1,0 +1,6 @@
+package constants
+
+const (
+	Gauge   = "gauge"
+	Counter = "counter"
+)
