+++ internal/middleware/middleware.go
@@ -1,0 +1,98 @@
+package middleware
+
+import (
+	"compress/gzip"
+	"fmt"
+	"io"
+	"net/http"
+)
+
+type compressWriter struct {
+	w  http.ResponseWriter
+	zw *gzip.Writer
+}
+
+func NewCompressWriter(w http.ResponseWriter) *compressWriter {
+	return &compressWriter{
+		w:  w,
+		zw: gzip.NewWriter(w),
+	}
+}
+
+func (c *compressWriter) Header() http.Header {
+	return c.w.Header()
+}
+
+func (c *compressWriter) Write(p []byte) (int, error) {
+	return c.zw.Write(p)
+}
+
+func (c *compressWriter) WriteHeader(statusCode int) {
+	if statusCode < 300 {
+		c.w.Header().Set("Content-Encoding", "gzip")
+	}
+	c.w.WriteHeader(statusCode)
+}
+
+func (c *compressWriter) Close() error {
+	return c.zw.Close()
+}
+
+type compressReader struct {
+	r  io.ReadCloser
+	zr *gzip.Reader
+}
+
+func NewCompressReader(r io.ReadCloser) (*compressReader, error) {
+	zr, err := gzip.NewReader(r)
+	if err != nil {
+		err = fmt.Errorf("ошибка 1: %s", err)
+		return nil, err
+	}
+
+	return &compressReader{
+		r:  r,
+		zr: zr,
+	}, nil
+}
+
+func (c compressReader) Read(p []byte) (n int, err error) {
+	return c.zr.Read(p)
+}
+
+func (c *compressReader) Close() error {
+	if err := c.r.Close(); err != nil {
+		err = fmt.Errorf("ошибка 2: %s", err)
+		return err
+	}
+	return c.zr.Close()
+}
+
+// func gzipMiddleware(h http.HandlerFunc) http.HandlerFunc {
+// 	return func(w http.ResponseWriter, r *http.Request) {
+// 		ow := w
+
+// 		supportsGzip := strings.Contains(r.Header.Get("Accept-Encoding"), "gzip")
+// 		supportsGzipJSON := strings.Contains(r.Header.Get("Accept"), "application/json")
+// 		supportsGzipHTML := strings.Contains(r.Header.Get("Accept"), "text/html")
+// 		if supportsGzip && (supportsGzipJSON || supportsGzipHTML) {
+// 			cw := newCompressWriter(w)
+// 			ow = cw
+// 			defer cw.Close()
+// 			ow.Header().Set("Content-Encoding", "gzip")
+// 		}
+
+// 		sendsGzip := strings.Contains(r.Header.Get("Content-Encoding"), "gzip")
+// 		if sendsGzip {
+// 			cr, err := newCompressReader(r.Body)
+// 			if err != nil {
+// 				w.WriteHeader(http.StatusInternalServerError)
+// 				return
+// 			}
+// 			r.Body = cr
+// 			defer cr.Close()
+// 		}
+
+// 		h.ServeHTTP(ow, r)
+// 	}
+// }
